refactor[#232] : login form mutationOption 리팩토링
