+++ src/features/auth/ui/LoginForm.stories.tsx
@@ -183,16 +183,21 @@ export const APIFailedTest: StoryObj<typeof LoginForm> = {
     msw: {
       handlers: [
         http.post("http://localhost/login", () => {
-          return HttpResponse.json({
-            code: 401,
-            message: "아이디 또는 비밀번호를 다시 확인해 주세요",
-            content: {
-              authorization: null,
-              role: null,
-              nickname: null,
-              userId: null,
+          return HttpResponse.json(
+            {
+              code: 401,
+              message: "아이디 또는 비밀번호를 다시 확인해 주세요",
+              content: {
+                authorization: null,
+                role: null,
+                nickname: null,
+                userId: null,
+              },
             },
-          });
+            {
+              status: 401,
+            },
+          );
         }),
       ],
     },
@@ -211,13 +216,6 @@ export const APIFailedTest: StoryObj<typeof LoginForm> = {
     });
 
     await step("API 요청이 실패한 경우엔 상태가 변경되지 않는다.", async () => {
-      // window.alert 목업하기
-      const originalAlert = window.alert;
-      let alertMessage;
-      window.alert = (message) => {
-        alertMessage = message;
-      };
-
       await userEvent.type($input, "abcd123@naver.com");
       await userEvent.type($password, "password");
       await userEvent.click($submit);
@@ -228,9 +226,6 @@ export const APIFailedTest: StoryObj<typeof LoginForm> = {
       expect(token).toBe(null);
       expect(role).toBe(null);
       expect(nickname).toBe(null);
-
-      expect(alertMessage).toEqual("아이디 또는 비밀번호를 다시 확인해 주세요");
-      window.alert = originalAlert;
     });
   },
 };
